requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.3
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drfdocs==0.0.11
gunicorn==19.7.1
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
Pygments==2.2.0
PyJWT==1.5.2
pytz==2017.2
requests==2.18.1
uritemplate==3.0.0
urllib3==1.21.1
whitenoise==3.3.0
